requirements.txt
flask==3.1.0
twilio==9.4.0
aiohttp==3.11.11
gunicorn==23.0.0
python-dotenv==1.0.1
